refactor: RootController 분리
